requirements.txt
duckduckgo-search
trafilatura
requests
pydantic
tenacity
python-dotenv
openai
requests